changed to CE_loss

+++ model_utils.py
@@ -95,7 +95,7 @@ class CXRImageDataset(Dataset):
         label_raw = label[0]
         label = convert_to_onehot(label[0])
         label = torch.tensor(label, dtype=torch.float32)
-        label_raw = torch.tensor(label_raw, dtype=torch.int)
+        label_raw = torch.tensor(label_raw, dtype=torch.long)
 
         sample = [image, label, label_raw]
